shiftedblog/pyproject.toml
[project]
name = "shiftedblog"
version = "0.1.0"
description = "Automated blogging tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.2",
    "django-axes>=8.0.0",
    "django-ckeditor-5>=0.2.18",
    "django-sslserver>=0.22",
    "django-taggit>=6.1.0",
    "django-two-factor-auth>=1.17.0",
    "phonenumbers>=9.0.8",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "qrcode>=7.4.2",
]
